refactor(models): use single pool for all operations
[Delivers #166168820]
